Separate feature in want/provide bits
We can support a wide array of asymmetric features by using two sets of
feature bits (`want_XXX` and `provide_XXX`).

This gives a lot of flexibility for use-cases where nodes may want to be both
a consumer and provider of a feature, and lets consumers require their peers
to be providers of the features they need.