BOLT 12: change `blindedpay` to one per *path*, not *hop*.
Bastien points out that a single value which applies across all hops
in the path is less revealing than per-hop information, and it's
simpler.  In practice, it can be argued that any extra cost would be
an issue for the recipient (they would accept less, for privacy) if it
turns out to be signficant.

Also makes the `blinded_capacities` language match.

Suggested-by: @t-bast
Signed-off-by: Rusty Russell <rusty@rustcorp.com.au>